Move Blueprint bundle editor into components package
