+++ packages/playground/components/package.json
@@ -30,6 +30,10 @@
 		"node": ">=20.10.0",
 		"npm": ">=10.2.3"
 	},
+	"dependencies": {
+		"@wp-playground/blueprints": "*",
+		"@wp-playground/storage": "*"
+	},
 	"peerDependencies": {
 		"@codemirror/autocomplete": "^6.18.6",
 		"@codemirror/commands": "^6.8.0",
