+++ packages/playground/components/src/BlueprintBundleEditor/index.ts
@@ -1,0 +1,6 @@
+export {
+	BlueprintBundleEditor,
+	type BlueprintBundleEditorHandle,
+	type BlueprintBundleEditorProps,
+} from './BlueprintBundleEditor';
+export { JSONSchemaEditor } from './json-schema-editor/json-schema-editor';
