+++ packages/playground/components/src/BlueprintBundleEditor/string-editor-modal.tsx
@@ -25,9 +25,15 @@ import {
 	keymap,
 	lineNumbers,
 } from '@codemirror/view';
-import { Button, Flex, FlexItem, SelectControl } from '@wordpress/components';
+import {
+	Button,
+	Flex,
+	FlexItem,
+	Modal,
+	SelectControl,
+} from '@wordpress/components';
+import React from 'react';
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { Modal } from '../modal';
 import {
 	languageLabels,
 	type SupportedLanguage,
@@ -210,9 +216,9 @@ export function StringEditorModal({
 		<Modal
 			title="Edit String"
 			onRequestClose={onClose}
-			className={css.stringEditorModal}
+			className={css['stringEditorModal']}
 		>
-			<div className={css.languageSelector}>
+			<div className={css['languageSelector']}>
 				<SelectControl
 					label="Language"
 					value={language}
@@ -223,8 +229,8 @@ export function StringEditorModal({
 					__nextHasNoMarginBottom
 				/>
 			</div>
-			<div className={css.editorContainer} ref={editorRef} />
-			<Flex justify="flex-end" className={css.buttons}>
+			<div className={css['editorContainer']} ref={editorRef} />
+			<Flex justify="flex-end" className={css['buttons']}>
 				<FlexItem>
 					<Button variant="tertiary" onClick={onClose}>
 						Cancel
