+++ packages/playground/components/src/index.ts
@@ -3,3 +3,4 @@ export * from './FilePickerTree/index';
 export * from './FilePickerControl/index';
 export * from './BinaryFilePreview';
 export * from './PlaygroundFileEditor/index';
+export * from './BlueprintBundleEditor/index';
