+++ packages/playground/website/src/components/blueprint-editor/BlueprintAuthoringEditor.ts
@@ -1,0 +1,1 @@
+export { JSONSchemaEditor } from '@wp-playground/components';
