+++ packages/playground/website/src/components/blueprint-editor/SiteBlueprintBundleEditor.spec.ts
@@ -14,8 +14,12 @@ import {
 	vi,
 } from 'vitest';
 import { InMemoryFilesystemBackend } from '@wp-playground/storage';
+import type { EventedFilesystem } from '@wp-playground/storage';
+import type * as BlueprintsModule from '@wp-playground/blueprints';
 import type * as BlueprintBundleStorageModule from '../../lib/state/opfs/opfs-blueprint-bundle-storage';
+import type * as SliceClientsModule from '../../lib/state/redux/slice-clients';
 import type * as SliceSitesModule from '../../lib/state/redux/slice-sites';
+import type * as SliceUiModule from '../../lib/state/redux/slice-ui';
 import {
 	readSiteBlueprintJson,
 	SiteBlueprintBundleEditor,
@@ -24,9 +28,28 @@ import {
 type SiteInfo = SliceSitesModule.SiteInfo;
 
 const mocks = vi.hoisted(() => ({
+	clientEntities: {} as Record<
+		string,
+		{ opfsSync?: { status: 'syncing' | 'error' } }
+	>,
+	componentProps: undefined as Record<string, unknown> | undefined,
+	createStoredSite: vi.fn(),
 	dispatch: vi.fn(),
 	persistBlueprintBundle: vi.fn(),
+	pruneAutosavedSites: vi.fn(),
+	removeClientInfo: vi.fn(),
+	removeSite: vi.fn(),
+	resolveRuntimeConfiguration: vi.fn(),
+	setActiveSite: vi.fn(),
+	setDockOperationNotice: vi.fn(),
+	setDockPaneOpen: vi.fn(),
+	updateSite: vi.fn(),
 	updateSiteMetadata: vi.fn(),
+}));
+
+vi.mock('@wp-playground/blueprints', async (importOriginal) => ({
+	...(await importOriginal<typeof BlueprintsModule>()),
+	resolveRuntimeConfiguration: mocks.resolveRuntimeConfiguration,
 }));
 
 vi.mock(
@@ -39,18 +62,42 @@ vi.mock(
 
 vi.mock('../../lib/state/redux/store', () => ({
 	useAppDispatch: () => mocks.dispatch,
-	useAppSelector: vi.fn(),
+	useAppSelector: (
+		selector: (state: {
+			clients: { entities: typeof mocks.clientEntities };
+		}) => unknown
+	) => selector({ clients: { entities: mocks.clientEntities } }),
+	setActiveSite: mocks.setActiveSite,
 }));
 
 vi.mock('../../lib/state/redux/slice-sites', async (importOriginal) => ({
 	...(await importOriginal<typeof SliceSitesModule>()),
+	createStoredSite: mocks.createStoredSite,
+	pruneAutosavedSites: mocks.pruneAutosavedSites,
+	removeSite: mocks.removeSite,
+	updateSite: mocks.updateSite,
 	updateSiteMetadata: mocks.updateSiteMetadata,
 }));
 
-vi.mock('./BlueprintBundleEditor', async () => {
+vi.mock('../../lib/state/redux/slice-clients', async (importOriginal) => ({
+	...(await importOriginal<typeof SliceClientsModule>()),
+	removeClientInfo: mocks.removeClientInfo,
+}));
+
+vi.mock('../../lib/state/redux/slice-ui', async (importOriginal) => ({
+	...(await importOriginal<typeof SliceUiModule>()),
+	setDockOperationNotice: mocks.setDockOperationNotice,
+	setDockPaneOpen: mocks.setDockPaneOpen,
+}));
+
+vi.mock('@wp-playground/components', async () => {
 	const { forwardRef } = await import('react');
 	return {
-		BlueprintBundleEditor: forwardRef(function BlueprintBundleEditorMock() {
+		BlueprintBundleEditor: forwardRef(function BlueprintBundleEditorMock(
+			props: Record<string, unknown>,
+			_ref
+		) {
+			mocks.componentProps = props;
 			return null;
 		}),
 	};
@@ -92,8 +139,23 @@ describe('SiteBlueprintBundleEditor', () => {
 		container = document.createElement('div');
 		document.body.append(container);
 		root = createRoot(container);
+		mocks.clientEntities = {};
+		mocks.componentProps = undefined;
+		mocks.createStoredSite.mockReset();
 		mocks.dispatch.mockReset();
 		mocks.persistBlueprintBundle.mockReset();
+		mocks.pruneAutosavedSites.mockReset();
+		mocks.removeClientInfo.mockReset();
+		mocks.removeSite.mockReset();
+		mocks.resolveRuntimeConfiguration.mockReset();
+		mocks.setActiveSite.mockReset();
+		mocks.setActiveSite.mockImplementation((slug) => ({
+			type: 'set-active-site',
+			payload: slug,
+		}));
+		mocks.setDockOperationNotice.mockReset();
+		mocks.setDockPaneOpen.mockReset();
+		mocks.updateSite.mockReset();
 		mocks.updateSiteMetadata.mockReset();
 	});
 
@@ -159,4 +221,190 @@ describe('SiteBlueprintBundleEditor', () => {
 		expect(mocks.dispatch).toHaveBeenCalledOnce();
 		expect(mocks.dispatch).toHaveBeenCalledWith(metadataAction);
 	});
+
+	it('runs an autosaved Blueprint in a new Playground', async () => {
+		const sourceSite = createStoredSite('source-site');
+		const newSite = createStoredSite('blueprint-copy');
+		const createAction = { type: 'create-stored-site' };
+		const pruneAction = { type: 'prune-autosaves' };
+		mocks.createStoredSite.mockReturnValue(createAction);
+		mocks.pruneAutosavedSites.mockReturnValue(pruneAction);
+		mocks.dispatch.mockImplementation((action) =>
+			action === createAction ? Promise.resolve(newSite) : action
+		);
+
+		const { filesystem, onPreview } = await renderEditor(sourceSite);
+		await act(async () => onPreview(filesystem));
+
+		expect(mocks.createStoredSite).toHaveBeenCalledWith(
+			sourceSite.metadata.name,
+			filesystem.backend,
+			undefined,
+			{
+				persistence: 'autosave',
+				siteSlugToReturnToIfBlueprintFails: sourceSite.slug,
+			}
+		);
+		expect(mocks.setActiveSite).toHaveBeenCalledWith(newSite.slug);
+		expect(mocks.pruneAutosavedSites).toHaveBeenCalledWith({
+			excludeSlugs: [sourceSite.slug, newSite.slug],
+		});
+	});
+
+	it('keeps a temporary Blueprint in the current Playground', async () => {
+		const site = {
+			slug: 'temporary-site',
+			metadata: {
+				name: 'Temporary Playground',
+				storage: 'none',
+				originalBlueprint: null,
+			},
+		} as SiteInfo;
+		const runtimeConfiguration = {
+			phpVersion: '8.3',
+			wpVersion: 'latest',
+		};
+		const updateAction = { type: 'update-site' };
+		mocks.resolveRuntimeConfiguration.mockResolvedValue(
+			runtimeConfiguration
+		);
+		mocks.updateSite.mockReturnValue(updateAction);
+		mocks.dispatch.mockImplementation((action) => action);
+
+		const { filesystem, onPreview } = await renderEditor(site);
+		await act(async () => onPreview(filesystem));
+
+		expect(mocks.createStoredSite).not.toHaveBeenCalled();
+		expect(mocks.resolveRuntimeConfiguration).toHaveBeenCalledWith(
+			filesystem
+		);
+		expect(mocks.updateSite).toHaveBeenCalledWith({
+			slug: site.slug,
+			changes: {
+				metadata: {
+					originalBlueprintSource: { type: 'none' },
+					originalBlueprint: filesystem,
+					runtimeConfiguration,
+					whenCreated: expect.any(Number),
+				},
+				originalUrlParams: undefined,
+			},
+		});
+	});
+
+	it('removes a failed run before pruning autosaves on retry', async () => {
+		const failedRun = createStoredSite('failed-run');
+		failedRun.metadata.siteSlugToReturnToIfBlueprintFails = 'original-site';
+		const replacement = createStoredSite('replacement');
+		const createAction = { type: 'create-stored-site' };
+		const removeAction = { type: 'remove-site' };
+		const pruneAction = { type: 'prune-autosaves' };
+		mocks.createStoredSite.mockReturnValue(createAction);
+		mocks.removeSite.mockReturnValue(removeAction);
+		mocks.pruneAutosavedSites.mockReturnValue(pruneAction);
+		mocks.dispatch.mockImplementation((action) =>
+			action === createAction ? Promise.resolve(replacement) : action
+		);
+
+		const { filesystem, onPreview } = await renderEditor(failedRun);
+		await act(async () => onPreview(filesystem));
+
+		expect(mocks.removeSite).toHaveBeenCalledWith(failedRun.slug);
+		expect(mocks.pruneAutosavedSites).toHaveBeenCalledWith({
+			excludeSlugs: ['original-site', replacement.slug],
+		});
+		const actions = mocks.dispatch.mock.calls.map(([action]) => action);
+		expect(actions.indexOf(removeAction)).toBeLessThan(
+			actions.indexOf(pruneAction)
+		);
+	});
+
+	it('waits for the source Playground to finish copying before running', async () => {
+		const sourceSite = createStoredSite('source-site');
+		const newSite = createStoredSite('blueprint-copy');
+		const createAction = { type: 'create-stored-site' };
+		mocks.clientEntities[sourceSite.slug] = {
+			opfsSync: { status: 'syncing' },
+		};
+		mocks.createStoredSite.mockReturnValue(createAction);
+		mocks.pruneAutosavedSites.mockReturnValue({ type: 'prune-autosaves' });
+		mocks.dispatch.mockImplementation((action) =>
+			action === createAction ? Promise.resolve(newSite) : action
+		);
+
+		const { filesystem, onPreview } = await renderEditor(sourceSite);
+		let preview!: Promise<void>;
+		act(() => {
+			preview = onPreview(filesystem);
+		});
+		await act(async () => Promise.resolve());
+		expect(mocks.createStoredSite).not.toHaveBeenCalled();
+
+		mocks.clientEntities[sourceSite.slug] = {};
+		await act(async () => {
+			root.render(
+				createElement(SiteBlueprintBundleEditor, { site: sourceSite })
+			);
+			await Promise.resolve();
+		});
+		await act(async () => preview);
+
+		expect(mocks.createStoredSite).toHaveBeenCalledOnce();
+	});
+
+	it('allows another run after Playground creation fails', async () => {
+		const sourceSite = createStoredSite('source-site');
+		const newSite = createStoredSite('blueprint-copy');
+		const createAction = { type: 'create-stored-site' };
+		let attempts = 0;
+		mocks.createStoredSite.mockReturnValue(createAction);
+		mocks.pruneAutosavedSites.mockReturnValue({ type: 'prune-autosaves' });
+		mocks.dispatch.mockImplementation((action) => {
+			if (action === createAction) {
+				attempts++;
+				return attempts === 1
+					? Promise.reject(new Error('Could not create site'))
+					: Promise.resolve(newSite);
+			}
+			return action;
+		});
+
+		const { filesystem, onPreview } = await renderEditor(sourceSite);
+		await expect(onPreview(filesystem)).rejects.toThrow(
+			'Could not create site'
+		);
+		await act(async () => onPreview(filesystem));
+
+		expect(mocks.createStoredSite).toHaveBeenCalledTimes(2);
+		expect(mocks.setActiveSite).toHaveBeenCalledWith(newSite.slug);
+	});
+
+	async function renderEditor(site: SiteInfo): Promise<{
+		filesystem: EventedFilesystem;
+		onPreview: (filesystem: EventedFilesystem) => Promise<void>;
+	}> {
+		await act(async () => {
+			root.render(createElement(SiteBlueprintBundleEditor, { site }));
+			await Promise.resolve();
+		});
+		await vi.waitFor(() => expect(mocks.componentProps).toBeDefined());
+		return mocks.componentProps as {
+			filesystem: EventedFilesystem;
+			onPreview: (filesystem: EventedFilesystem) => Promise<void>;
+		};
+	}
+
+	function createStoredSite(slug: string): SiteInfo {
+		return {
+			slug,
+			metadata: {
+				id: slug,
+				name: 'Stored Blueprint',
+				storage: 'opfs',
+				persistence: 'autosave',
+				originalBlueprint: new InMemoryFilesystemBackend(),
+				originalBlueprintSource: { type: 'opfs-site' },
+			},
+		} as SiteInfo;
+	}
 });
